mgr/cephadm: add options for haproxy to verify backend ssl cert
This requires us to have the CA cert used to sign that
certificate. In the case where we generate our own cert
for the ingress service, that's just our cephadm CA cert.
Otherwise, we'll need the user to pass in the CA cert
themselves for us to use for verification.

Signed-off-by: Adam King <adking@redhat.com>